+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add app::version_str().
 - Add dialog::message_title() and message_title_default().
 - Shift experimental Flow widget to its own [crate](https://github.com/MoAlyousef/fltk-flow).
+- Add TreeItem::draw_item_content.
 
 ## [1.2.17] - 2021-11-21
 - Add draw::draw_check().